feat: add UDS byte encoding, service lookup, and interactive byte grid
Integrate CDA EcuManager for UDS translation: service lookup by SID/name,
byte-pattern extraction from request/response DOPs, and parameter encoding.

Frontend additions:
- ByteGridView: hex byte grid with field highlighting and selection
- DetailPane: UDS encoding panel with editable parameters and live preview
- UdsServiceTranslator: standalone service translation component
- Settings: CDA base URL configuration

Backend additions:
- UdsTranslator wrapping CDA EcuManager with async init and variant support
- byte_pattern module for extracting const/param byte layouts from DOPs
- Tauri commands: uds_load, uds_lookup, uds_to_sovd, sovd_to_uds, service_schema
- Navigate-to now auto-expands target node children

Remove accidentally committed generated schema files (gen/schemas/).

+++ src/mdd-core/src/diff/compare.rs
@@ -386,7 +386,7 @@ fn compare_services(
         });
     }
     if old.request_sub_function != new.request_sub_function {
-        let format_sub_fn = |opt: Option<(u16, u32)>| -> String {
+        let format_sub_fn = |opt: Option<(u32, u32)>| -> String {
             opt.map_or_else(
                 || "None".to_owned(),
                 |(sub_fn, bit_len)| format!("0x{sub_fn:X} ({bit_len} bits)"),
